CPU-separation: move `cpu_affinity.cpp` to common code
We would need this level of CPU manipulation with other devices as well,
for things like NUMA locality.
