+++ framework/device/cpu/sysdeps/darwin/meson.build
@@ -3,6 +3,5 @@
 
 framework_files += \
     files(
-        'cpu_affinity.cpp',
         '../generic' / 'physicaladdress.c',
     )
