+++ framework/sysdeps/freebsd/cpu_affinity.cpp
