+++ framework/sysdeps/freebsd/meson.build
@@ -3,6 +3,7 @@
 
 framework_files += \
     files(
+        'cpu_affinity.cpp',
         'malloc.cpp',
         '../generic/kvm.c',
         '../generic/memfpt.c',
