+++ framework/topology.h
@@ -171,6 +171,7 @@ private:
 
 LogicalProcessorSet ambient_logical_processor_set();
 bool pin_to_logical_processor(LogicalProcessor, const char *thread_name = nullptr);
+bool pin_thread_to_logical_processor(LogicalProcessor n, tid_t thread_id, const char *thread_name = nullptr);
 bool pin_to_logical_processors(DeviceRange, const char *thread_name);
 
 void apply_deviceset_param(char *param);
